requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# 数据处理
numpy>=1.21.0
opencv-python>=4.8.0
Pillow>=9.0.0

# 评估工具
pycocotools>=2.0.6
scikit-learn>=1.0.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 工具库
tqdm>=4.64.0
PyYAML>=6.0

# GPU支持
# cudatoolkit>=11.8  # 通常通过conda安装